Replace main function with a symbol indexer
This program runs in two modes, either a single-file mode (which outputs
on stdout), or batch mode, (which outputs to a file based on the first
argument and an environment variable).

In either case, the program reads sepecified .blob files, iterates over
all symbols contained in the index, and outputs TSV data for those
symbols which are defined by the file being indexed.

The resulting output can be used as an index to search for the
definition of any symbol, without having an LSP-enabled editor open at
the time. A commit in my dotfiles repo (brenns10/dotfiles@79ade3b758a4
"tagsearch: add a system for searching indexed symbols") demonstrates
the use of this system.